requirements.txt
charset-normalizer==2.0.12
Flask==2.0.2
Flask-Session==0.4.0
tensorflow-cpu==2.7.0
gensim==4.1.2
gunicorn==20.1.0
Jinja2==3.1.1
keras==2.7.0
requests==2.27.1
langid==1.1.6
nltk==3.7
numpy==1.19.5
PyMySQL==1.0.2